gpu: nvgpu: pmu: fix MISRA 10.3 violations
This fixes a number of MISRA 10.3 violations in the common/pmu/pmu_g*
files.  MISRA Rule 10.3 prohibits implicit assignment of different size
or essential types.

JIRA NVGPU-1008

Reviewed-on: https://git-master.nvidia.com/r/2017608
GVS: Gerrit_Virtual_Submit
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>

+++ drivers/gpu/nvgpu/include/nvgpu/gk20a.h
@@ -1813,7 +1813,7 @@ struct gk20a {
 	struct gpu_ops ops;
 	u32 mc_intr_mask_restore[4];
 	/*used for change of enum zbc update cmd id from ver 0 to ver1*/
-	u32 pmu_ver_cmd_id_zbc_table_update;
+	u8 pmu_ver_cmd_id_zbc_table_update;
 	bool pmu_lsf_pmu_wpr_init_done;
 	u32 pmu_lsf_loaded_falcon_id;
